--- Assets/_BForBoss/Scripts/InputSettingsViewBehaviour.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b0b6b40ef2765472eb92aa5da5b8e534
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
