+++ Assets/_BForBoss/_Leaderboard/Scripts/Username.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b16cf479e237345fcb342b0a7ca4eaff
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
