+++ Assets/_BForBoss/_Leaderboard/Scripts/Username/ForcedInputUsernameViewBehaviour.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1cf4310bd47d14dfe84f98d8dfe1d648
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
