+++ Assets/_BForBoss/_Leaderboard/Scripts/Username/InputUsername.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a10b2a70078b149a48e50ec6c311c685
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
