+++ Assets/_BForBoss/_Leaderboard/Scripts/Username/InputUsernameViewModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4f865147776c7438f8f344a131df9b8b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
