+++ Assets/_BForBoss/_Leaderboard/UsernameInputPanel.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 231a2d720be814ae2b95e84d71b8b863
+guid: f1d69b30a41d14e169372bc26aeed53b
 PrefabImporter:
   externalObjects: {}
   userData:
